Refactor application structure by separating AppShell and AppRoutes into their own files; Implement lazy loading for routes and improve code organization; Update shared components and fix tab handling.
