+++ react/src/components/shared.tsx
