+++ react/src/pages/Listagem.tsx
@@ -2,49 +2,23 @@ import { useMemo, useState } from 'react'
 import { Dropdown, Modal, Nav, Tab } from 'react-bootstrap'
 import { LIST_FILTERS, LIST_RECORDS } from '../data/examples'
 import type { ListFilter } from '../data/examples'
-import { LoadingOverlay, MdiIcon, PageHeader } from './shared'
+import { ItemActions } from '../components/ItemActions'
+import { LoadingOverlay, MdiIcon, PageHeader } from '../components/shared'
 
 type ModalProps = {
   show: boolean
   onHide: () => void
 }
 
-function RecordActions({ id }: { id: number }) {
-  return (
-    <Dropdown align="end" className="d-inline-block">
-      <Dropdown.Toggle
-        variant="link"
-        className="no-caret"
-        id={`registro-acoes-${id}`}
-        aria-label="Abrir ações do registro"
-      >
-        <MdiIcon name="dots-vertical" />
-      </Dropdown.Toggle>
-      <Dropdown.Menu>
-        <Dropdown.Item as="button" type="button">
-          <MdiIcon name="pencil" className="me-2" />
-          Editar
-        </Dropdown.Item>
-        <Dropdown.Divider />
-        <Dropdown.Item as="button" type="button">
-          <MdiIcon name="trash-can-outline" className="me-2" />
-          Excluir
-        </Dropdown.Item>
-      </Dropdown.Menu>
-    </Dropdown>
-  )
-}
-
 function RegisterModal({ show, onHide }: ModalProps) {
-  const [activeStep, setActiveStep] = useState('1')
-  const stepNumber = Number(activeStep)
+  const [activeStep, setActiveStep] = useState(1)
 
   const close = () => {
-    setActiveStep('1')
+    setActiveStep(1)
     onHide()
   }
 
-  const next = () => setActiveStep(String(Math.min(stepNumber + 1, 3)))
+  const next = () => setActiveStep(Math.min(activeStep + 1, 3))
 
   return (
     <Modal show={show} onHide={close} size="lg">
@@ -56,32 +30,32 @@ function RegisterModal({ show, onHide }: ModalProps) {
         <Modal.Body>
           <div className="row">
             <div className="col-12">
-              <Tab.Container activeKey={activeStep} onSelect={(key) => key && setActiveStep(key)}>
+              <Tab.Container activeKey={activeStep} onSelect={(key) => key && setActiveStep(Number(key))}>
                 <Nav variant="tabs" className="nav-wizard">
                   <Nav.Item>
-                    <Nav.Link eventKey="1" className="checked">
+                    <Nav.Link eventKey={1} className="checked">
                       Etapa 1
                     </Nav.Link>
                   </Nav.Item>
                   <Nav.Item>
-                    <Nav.Link eventKey="2" className="checked">
+                    <Nav.Link eventKey={2} className="checked">
                       Etapa 2
                     </Nav.Link>
                   </Nav.Item>
                   <Nav.Item>
-                    <Nav.Link eventKey="3">
+                    <Nav.Link eventKey={3}>
                       Etapa 3
                     </Nav.Link>
                   </Nav.Item>
                   <Nav.Item>
-                    <Nav.Link eventKey="4" disabled>
+                    <Nav.Link eventKey={4} disabled>
                       Etapa 4
                     </Nav.Link>
                   </Nav.Item>
                 </Nav>
 
                 <Tab.Content className="mt-3">
-                  <Tab.Pane eventKey="1">
+                  <Tab.Pane eventKey={1}>
                     <div className="row g-3">
                       <div className="col-12">
                         <label htmlFor="modalNome" className="form-label required">
@@ -110,13 +84,13 @@ function RegisterModal({ show, onHide }: ModalProps) {
                       </div>
                     </div>
                   </Tab.Pane>
-                  <Tab.Pane eventKey="2">
+                  <Tab.Pane eventKey={2}>
                     <p className="mb-0">Conteúdo da etapa 2</p>
                   </Tab.Pane>
-                  <Tab.Pane eventKey="3">
+                  <Tab.Pane eventKey={3}>
                     <p className="mb-0">Conteúdo da etapa 3</p>
                   </Tab.Pane>
-                  <Tab.Pane eventKey="4">
+                  <Tab.Pane eventKey={4}>
                     <p className="mb-0">Conteúdo da etapa 4</p>
                   </Tab.Pane>
                 </Tab.Content>
@@ -134,7 +108,7 @@ function RegisterModal({ show, onHide }: ModalProps) {
               </button>
             </div>
             <div className="col-auto">
-              <button type="button" className="btn btn-primary" disabled={stepNumber !== 3}>
+              <button type="button" className="btn btn-primary" disabled={activeStep !== 3}>
                 Salvar
               </button>
               <button type="button" className="btn btn-secondary ms-1" onClick={close}>
@@ -336,7 +310,7 @@ export function Listagem() {
                     <span className={record.badgeClass}>{record.status}</span>
                   </td>
                   <td className="text-end">
-                    <RecordActions id={record.id} />
+                    <ItemActions id={record.id} ariaLabel="Abrir ações do registro" />
                   </td>
                 </tr>
               ))}
